+++ src/messages/MessageElement.hpp
@@ -508,6 +508,7 @@ public:
                         const MessageLayoutContext &ctx) override;
 
     TextElement *formatTime(const QTime &time);
+    MessageElement *setLink(const Link &link) override;
 
     QJsonObject toJson() const override;